Rename Rust test files for consistency
