+++ lambda_compiler/src/compilation_tests.rs
