+++ lambda_compiler/src/examples_tests.rs
