+++ lambda_compiler/src/lib.rs
@@ -3,28 +3,28 @@
 pub mod compilation;
 
 #[cfg(test)]
-mod compilation_test;
+mod compilation_tests;
 
 pub mod ast;
 pub mod type_checking;
 
 #[cfg(test)]
-mod type_checking_test;
+mod type_checking_tests;
 
 pub mod parsing;
 
 #[cfg(test)]
-mod parsing_test;
+mod parsing_tests;
 
 pub mod tokenization;
 
 #[cfg(test)]
-mod tokenization_test;
+mod tokenization_tests;
 
 #[cfg(test)]
-mod examples_test;
+mod examples_tests;
 
 pub mod format;
 
 #[cfg(test)]
-mod format_test;
+mod format_tests;
